+++ src/main/java/om/example/Sequence_measurement_api/repositries/SequenceHistory.java
@@ -1,4 +1,0 @@
-package om.example.Sequence_measurement_api.repositries;
-
-public interface SequenceHistory {
-}