+++ ChangeLog.md
@@ -1,4 +1,10 @@
 # Change Log
+## Version[0.0.4] - 22-05-2026
+### Fixed
+- Added `SequenceRepo` interface for DB connection.
+- Established a connection with the DB and tested with multiple requests.
+- PUT request now works with raw JSON format to update input value.
+
 ## Version[0.0.3] - 21-05-2026
 ### Fixed
 - Added PUT request to edit using input with endpoint `/convert-measurements/{id}`
@@ -10,7 +16,7 @@
 - User input is converted to lower case.
 - Handled invalid user input (numbers and special characters).
 
-## Version[0.0.1] - 21-05-2026
+## Version[0.0.1] - 20-05-2026
 ### Added
 - Initial release of Sequence Measurement API.
 - Core algorithm for decoding measurement sequences ("SequenceService") with z-multiplier encoding.